--- backend/movie_application_backend/src/main/java/com/group_1/movie_application_backend/MovieApplicationBackendApplication.java
@@ -1,15 +1,0 @@
-package com.group_1.movie_application_backend;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MovieApplicationBackendApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(MovieApplicationBackendApplication.class, args);
-		
-	}
-
-}
-
