--- backend/movie_application_backend/src/main/java/com/group_1/movie_application_backend/controllers/EventController.java
@@ -1,9 +1,0 @@
-package com.group_1.movie_application_backend.controllers;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/event")
-public class EventController {
-
-}
