--- backend/movie_application_backend/src/main/java/com/group_1/movie_application_backend/controllers/UserController.java
@@ -1,15 +1,0 @@
-package com.group_1.movie_application_backend.controllers;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/user")
-public class UserController {
-    // This is a placeholder for the TMDBController class
-    // This class will be used to handle requests to the TMDB API
-    // The TMDB API will be used to get information about movies and TV shows
-    // The information will be used to display the movies and TV shows on the website
-    // The information will also be used to display the details of a movie or TV show when a user clicks on it
-    
-}
