--- backend/movie_application_backend/src/main/java/com/group_1/movie_application_backend/models/Group.java
@@ -1,5 +1,0 @@
-package com.group_1.movie_application_backend.models;
-
-public class Group {
-
-}
