--- backend/movie_application_backend/src/main/java/com/group_1/movie_application_backend/models/Movie.java
@@ -1,10 +1,0 @@
-package com.group_1.movie_application_backend.models;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
-
-@Service
-public class Movie {
-    @Value("${tmdb.api.key}")
-    private String tmdbApiKey;
-}
