--- backend/movie_application_backend/src/main/java/com/group_1/movie_application_backend/models/User.java
@@ -1,15 +1,0 @@
-package com.group_1.movie_application_backend.models;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-
-@Entity
-public class User {
-    @Id
-    @GeneratedValue(strategy=GenerationType.AUTO)
-    private Long userId;
-    @
-    
-}
