--- backend/movie_application_backend/src/main/java/com/group_1/movie_application_backend/repositories/EventRepository.java
@@ -1,5 +1,0 @@
-package com.group_1.movie_application_backend.repositories;
-
-public class EventRepository {
-    
-}
