--- backend/movie_application_backend/src/main/resources/application.properties
@@ -1,2 +1,0 @@
-spring.application.name=movie_application_backend
-
