--- backend/movie_application_backend/src/test/java/com/group_1/movie_application_backend/MovieApplicationBackendApplicationTests.java
@@ -1,13 +1,0 @@
-package com.group_1.movie_application_backend;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MovieApplicationBackendApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
