+++ backend/pom.xml
@@ -63,6 +63,10 @@
 			<version>2.4.0</version>
 			<!--<scope>test</scope>-->
 		</dependency>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-data-jpa</artifactId>
+		</dependency>
 	</dependencies>
 
 	<build>
