+++ backend/src/main/java/fi/oamk/muuvi/backend/controller/EventController.java
@@ -1,0 +1,9 @@
+package fi.oamk.muuvi.backend.controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/event")
+public class EventController {
+
+}
