+++ backend/src/main/java/fi/oamk/muuvi/backend/models/Event.java
@@ -1,0 +1,5 @@
+package fi.oamk.muuvi.backend.models;
+
+public class Event {
+
+}
