+++ backend/src/main/java/fi/oamk/muuvi/backend/models/User.java
@@ -1,0 +1,15 @@
+package fi.oamk.muuvi.backend.models;
+
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+
+@Entity
+public class User {
+    @Id
+    @GeneratedValue(strategy=GenerationType.AUTO)
+    private Long userId;
+    
+}
