+++ backend/src/main/java/fi/oamk/muuvi/backend/repositories/EventRepository.java
@@ -1,0 +1,5 @@
+package fi.oamk.muuvi.backend.repositories;
+
+public class EventRepository {
+    
+}
