+++ backend/src/main/java/fi/oamk/muuvi/backend/repositories/ReviewRepository.java
@@ -1,0 +1,5 @@
+package fi.oamk.muuvi.backend.repositories;
+
+public class ReviewRepository {
+
+}
