+++ backend/src/main/java/fi/oamk/muuvi/backend/repositories/UserRepository.java
@@ -1,0 +1,5 @@
+package fi.oamk.muuvi.backend.repositories;
+
+public class UserRepository {
+
+}
