+++ backend/src/main/java/fi/oamk/muuvi/backend/services/UserService.java
@@ -1,0 +1,5 @@
+package fi.oamk.muuvi.backend.services;
+
+public class UserService {
+
+}
